+++ db/migrate/20251122160000_add_knowledge_base_to_sdr_ia_configs.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class AddKnowledgeBaseToSdrIaConfigs < ActiveRecord::Migration[7.0]
+  def change
+    add_column :sdr_ia_configs, :knowledge_base, :text, default: ''
+  end
+end